Client: the terrain chunk builder and the height and render views, registered (m0-foundation/T20, in progress)
The previous commit carried these files without their project entries.
HeightView and RenderView are the Core aggregates of TechnicalDesign.md
§6.3; TerrainChunk builds a 32-by-32-cell chunk at a stride of 1, 2, 4
or 8 with a skirt and colours every vertex by the Species formula into
a built-in palette; Lighting.h holds the Species light pair and specifies
the unlit team-colour slot for M1.

+++ Client/TerrainChunk.h
@@ -47,7 +47,7 @@ public:
   /// _u the slope term and _v the height term, both 0 to 1, clamped.
   [[nodiscard]] std::uint32_t Lookup(float _u, float _v) const noexcept;
 
-  std::array<std::uint32_t, SIDE * SIDE> colors{};
+  std::array<std::uint32_t, static_cast<std::size_t>(SIDE) * SIDE> colors{};
 };
 
 struct TerrainMesh
